Final_Work/requirements.txt
# Core dependencies
langchain>=0.2.3
langchain-openai>=0.1.7
pinecone>=3.0.0
scikit-learn>=1.5.0
sentence-transformers>=3.0.1
python-dotenv>=1.0.0

# Document processing
pdfplumber>=0.10.0
pandas>=2.0.0
openpyxl>=3.1.0

# AI and ML
openai>=1.0.0
numpy>=1.24.0
scipy>=1.10.0

# Evaluation
ragas>=0.2.0
datasets>=2.14.0

# LlamaExtract for enhanced data extraction
llama-cloud-services>=0.1.0
pydantic>=2.0.0